feat: add modal component as a wrapper
